inbound: move to direct h2 usage
